make player archiving/unarchiving vanilla style, and make ticcmd use 8 bit forward/sidemove again

+++ d_ticcmd.h
@@ -27,12 +27,12 @@
 // plus a checksum for internal state consistency.
 // 10 bytes each..
 typedef struct {
-	int16_t	forwardmove;	// *2048 for move
-	int16_t	sidemove;	// *2048 for move
-    int16_t	angleturn;	// <<16 for angle delta
-    int16_t	consistancy;	// checks for net game
-    byte	chatchar;
-    byte	buttons;
+	int8_t  forwardmove;	// *2048 for move
+	int8_t  sidemove;	// *2048 for move
+    int16_t angleturn;	// <<16 for angle delta
+    int16_t consistancy;	// checks for net game
+    byte    chatchar;
+    byte    buttons;
 } ticcmd_t;
